+++ project/src/components/film-card/film-card.tsx
