+++ project/src/pages/film-screen/film-screen.tsx
@@ -1,9 +1,9 @@
 import { Helmet } from 'react-helmet-async';
 import { Link, useParams } from 'react-router-dom';
-import FilmCard from '../../components/film-card';
-import Footer from '../../components/footer';
-import Header from '../../components/header';
-import UserBlock from '../../components/user-block';
+import FilmCard from '../../components/film-card/film-card';
+import Footer from '../../components/footer/footer';
+import Header from '../../components/header/header';
+import UserBlock from '../../components/user-block/user-block';
 import { FilmType } from '../../types/film-type';
 import NotFoundScreen from '../not-found-screen/not-found-screen';
