+++ project/src/pages/not-found-screen/not-found-screen.tsx
@@ -1,7 +1,7 @@
 import { Helmet } from 'react-helmet-async';
 import { Link } from 'react-router-dom';
-import Footer from '../../components/footer';
-import Header from '../../components/header';
+import Footer from '../../components/footer/footer';
+import Header from '../../components/header/header';
 
 export default function NotFoundScreen(): JSX.Element {
   return (
